requirement.txt
asn1crypto==0.22.0
bcrypt==3.1.3
cffi==1.10.0
cryptography==1.9
Django==1.8.2
djangorestframework==3.6.3
enum34==1.1.6
Fabric==1.13.2
fabtools==0.20.0
idna==2.5
ipaddress==1.0.18
MySQL-python==1.2.5
paramiko==2.2.1
pyasn1==0.2.3
pycparser==2.18
PyNaCl==1.1.2
six==1.10.0
